Factor out loginHandler
